Delete notice Header and add back to notice
